requirements.txt
celery>=4.4.2
datedelta>=1.3
datetime-truncate>=1.1.0
Django>=3.0.4
djangorestframework>=3.11.0
djangorestframework-simplejwt>=4.4.0
graphene-django>=2.9.0
graphene-django-extras>=0.4.8
graphene-file-upload>=1.2.2
django-graphql-jwt>=0.3.0
django-environ>=0.4.5
django-filter>=2.2.0
django-split-settings>=1.0.0
django-cors-headers>=3.2.1
gunicorn>=20.0.4
pandas>=1.0.3
psycopg2>=2.8.4
redis>=3.4.1